+++ Libraries/SideScroll/Tasks/TaskInstance.cs
@@ -410,7 +410,7 @@ public class TaskInstance : INotifyPropertyChanged
 		NotifyPropertyChanged(nameof(Finished));
 		NotifyPropertyChanged(nameof(CancelVisible));
 
-		Call.Log.Add("Finished", new Tag("Time", _stopwatch.ElapsedMilliseconds / 1000.0));
+		Call.Log.Add("Finished", new Tag("Duration", _stopwatch.Elapsed));
 
 		if (ParentTask == null)
 		{